+++ scripts/gen_docs.sh
@@ -15,6 +15,7 @@ EXAMPLES=(
     chord
     contour
     density
+    forest
     ridgeline
     dotplot
     heatmap